Joins Types
-- INSTRUCTIONS --

Complete the tasks in the Understanding MySQL joins document.  Copy and paste your join commands and results to this assignment. I would like for you to use your database on Google Cloud for this assignment.

Understanding SQL Joins.docx Download Understanding SQL Joins.docx

Note:  When a value of NULL is in the table, just do not enter a value for that attribute when adding the data.