Netflix parts 1-3
Creating a Netflix database part 1-3
-- INSTRUCTIONS --

(Part1)
Based on the NextFlix database design (NextFlix.pdf  Download NextFlix.pdf).  Use the appropriate SQL commands to:

1.) Create the database
2.) Create the tables in including appropriate data types (Actors, Movies, ActorMovies, Directors)
3.) Create appropriate foreign key constraints
4.) Add data including at least 5 movies, the director, and at least 3 actors from each movie (put yourself in as an actor in at least two movies)

5.) Create and run the following queries.  For each query copy and paste or screenshot the query and submit the result to assignment:

a query that joins the tables and displays all actors (firstname, lastname) and all movies (title)

a query that joins the tables and displays all actors in a particular movie

a query that joins the tables and displays only the movies in which you are an actor
a query that joins the tables and displays the director and movie title of the movies in which you are an actor
a query that joins the tables and displays the total salaries of the actors for each movie
6.) Submit your queries and the results they produce.

(PART 2)
Building on your NextFlix database, you will create a basic web application that can execute read operations on your database. You can use the index.php  Download index.phpfile as a template.

Your application must be able to successfully read from your database and display data from your Movies, Actors, and Directors tables.  Feel free to add functionality and improve the aesthetics of the application as your skills allow.

(PART 3)
Building on your NextFlix database, you will create a basic web application that can execute create, read, update, and delete (CRUD) operations on your database.

Your application must be able to successfully perform these operations on at least one table in your database. Feel free to add functionality and improve the aesthetics of the application as your skills allow.